Moved follow methods to AccountService
