+++ server/src/main/java/edu/sjsu/moth/server/controller/AccountController.java
@@ -13,6 +13,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RequestPart;
@@ -121,5 +122,6 @@ public class AccountController {
         // it's not clear what we need to do with the user...
         return accountService.getAccount(id).map(ResponseEntity::ok);
     }
+
 
 }
