+++ server/src/main/java/edu/sjsu/moth/server/controller/InboxController.java
@@ -12,6 +12,7 @@ import edu.sjsu.moth.server.db.Followers;
 import edu.sjsu.moth.server.db.FollowersRepository;
 import edu.sjsu.moth.server.db.Following;
 import edu.sjsu.moth.server.db.FollowingRepository;
+import edu.sjsu.moth.server.service.AccountService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -36,14 +37,9 @@ import static org.springframework.beans.support.PagedListHolder.DEFAULT_PAGE_SIZ
 
 @RestController
 public class InboxController {
-    @Autowired
-    FollowersRepository followersRepository;
 
     @Autowired
-    FollowingRepository followingRepository;
-
-    @Autowired
-    AccountRepository accountRepository;
+    AccountService accountService;
 
     //required to map payload from JSON to a Java Object for data access
     ObjectMapper mappedLoad;
@@ -57,34 +53,7 @@ public class InboxController {
         String requestType = inboxNode.get("type").asText();
         // follow or unfollow requests
         if (requestType.equals("Follow") || requestType.equals("Undo"))
-            return followerHandler(id, inboxNode, requestType);
-        return Mono.empty();
-    }
-
-    public Mono<String> followerHandler(String id, JsonNode inboxNode, String requestType) {
-        String follower = inboxNode.get("actor").asText();
-        if (requestType.equals("Follow")) {
-            // check id
-            if (accountRepository.findItemByAcct(id)==null) {
-                return Mono.error(new RuntimeException("Error: Account to follow doesn't exist."));
-            }
-            // find id, grab arraylist, append
-            else {
-                return followersRepository.findItemById(id)
-                        .switchIfEmpty(Mono.just(new Followers(id, new ArrayList<>())))
-                        .flatMap(followedUser -> {
-                            followedUser.getFollowers().add(follower);
-                            return followersRepository.save(followedUser).thenReturn("done");
-                        });
-            }
-        }
-        if (requestType.equals("Undo")) {
-            // find id, grab arraylist, remove
-            return followersRepository.findItemById(id).flatMap(followedUser -> {
-                followedUser.getFollowers().remove(follower);
-                return followersRepository.save(followedUser).thenReturn("done");
-            });
-        }
+            return accountService.followerHandler(id, inboxNode, requestType);
         return Mono.empty();
     }
 
@@ -92,53 +61,14 @@ public class InboxController {
     public Mono<UsersFollowResponse> usersFollowing(@PathVariable String id,
                                                     @RequestParam(required = false) Integer page,
                                                     @RequestParam(required = false) Integer limit) {
-        return usersFollow(id, page, limit, "following");
+        return accountService.usersFollow(id, page, limit, "following");
     }
 
     @GetMapping("/users/{id}/followers")
     public Mono<UsersFollowResponse> usersFollowers(@PathVariable String id,
                                                     @RequestParam(required = false) Integer page,
                                                     @RequestParam(required = false) Integer limit) {
-        return usersFollow(id, page, limit, "followers");
-    }
-
-    public Mono<UsersFollowResponse> usersFollow(String id, @RequestParam(required = false) Integer page,
-                                                 @RequestParam(required = false) Integer limit, String followType) {
-        var items = followType.equals("following") ? followingRepository.findItemById(id)
-                .map(Following::getFollowing) : followersRepository.findItemById(id).map(Followers::getFollowers);
-        String returnID = MothController.BASE_URL + "/users/" + id + followType;
-        int pageSize = limit != null ? limit : DEFAULT_PAGE_SIZE;
-        if (page == null) {
-            String first = returnID + "?page=1";
-            return items.map(
-                    v -> new UsersFollowResponse(returnID, "OrderedCollection", v.size(), first, null, null, null));
-        } else { // page number is given
-            int pageNum = page < 1 ? 1 : page;
-            return items.map(v -> {
-                String newReturnID = limit != null ? returnID + "?page=" + page + "&limit=" + limit : returnID +
-                        "?page=" + page;
-                if (pageNum * pageSize >= v.size()) { // no next page
-                    return new UsersFollowResponse(newReturnID, "OrderedCollectionPage", v.size(), null, null, returnID,
-                                                   paginateFollowers(v, pageNum, pageSize));
-                } else {
-                    String next = returnID + "?page=" + (pageNum + 1);
-                    if (limit != null) {
-                        next += "&limit=" + limit;
-                    }
-                    return new UsersFollowResponse(newReturnID, "OrderedCollectionPage", v.size(), null, next, returnID,
-                                                   paginateFollowers(v, pageNum, pageSize));
-                }
-            });
-        }
-    }
-
-    public List<String> paginateFollowers(ArrayList<String> followers, int pageNo, int pageSize) {
-        int startIndex = (pageNo - 1) * pageSize;
-        int endIndex = Math.min(startIndex + pageSize, followers.size());
-        if (startIndex >= followers.size()) {
-            return Collections.emptyList();
-        }
-        return followers.subList(startIndex, endIndex);
+        return accountService.usersFollow(id, page, limit, "followers");
     }
 
     @JsonInclude(JsonInclude.Include.NON_NULL)
